supabase/requirements.txt
supabase>=2.10.0
python-dotenv>=1.0.0
